refactor: update Octicon, Pagination tests to vitest
