+++ packages/react/jest.config.js
@@ -30,7 +30,9 @@ module.exports = {
     '<rootDir>/src/FeatureFlags/',
     '<rootDir>/src/FormControl/__tests__/FormControl.Validation.test.tsx',
     '<rootDir>/src/NavList/',
+    '<rootDir>/src/Octicon/',
     '<rootDir>/src/Pagehead/',
+    '<rootDir>/src/Pagination/',
     '<rootDir>/src/ProgressBar/',
     '<rootDir>/src/Radio/',
     '<rootDir>/src/RadioGroup/',
