+++ packages/react/src/Octicon/Octicon.test.tsx
@@ -1,22 +1,21 @@
 import {XIcon} from '@primer/octicons-react'
+import {render} from '@testing-library/react'
+import {describe, expect, it} from 'vitest'
 import Octicon from '../Octicon'
-import {behavesAsComponent, checkExports} from '../utils/testing'
-import {render as HTMLRender} from '@testing-library/react'
-import axe from 'axe-core'
 
 describe('Octicon', () => {
-  behavesAsComponent({
-    Component: Octicon,
-    toRender: () => <Octicon icon={XIcon} />,
+  it('should support `className` on the outermost element', () => {
+    const {container} = render(<Octicon icon={XIcon} className="test-class" />)
+    expect(container.firstChild).toHaveClass('test-class')
   })
 
-  checkExports('Octicon', {
-    default: Octicon,
+  it('should support `sx`', () => {
+    const {container} = render(<Octicon icon={XIcon} sx={{color: 'red'}} />)
+    expect(container.firstChild).toHaveStyle('color: rgb(255, 0, 0)')
   })
 
-  it('should have no axe violations', async () => {
-    const {container} = HTMLRender(<Octicon icon={XIcon} />)
-    const results = await axe.run(container)
-    expect(results).toHaveNoViolations()
+  it('should pass along props to the outermost element', () => {
+    const {container} = render(<Octicon icon={XIcon} data-testid="test-id" />)
+    expect(container.firstChild).toHaveAttribute('data-testid', 'test-id')
   })
 })
