+++ packages/react/src/Pagination/Pagination.features.stories.tsx
@@ -2,7 +2,7 @@ import {useState} from 'react'
 import type {Meta} from '@storybook/react'
 import type {ComponentProps} from '../utils/types'
 import Pagination from './Pagination'
-import {ReactRouterLikeLink} from '../__tests__/mocks/ReactRouterLink'
+import {ReactRouterLikeLink} from './mocks/ReactRouterLink'
 
 export default {
   title: 'Components/Pagination/Features',
